wip: split config and context modules and tests, rename types.monads
