+++ src/cdf/core/constants.py
@@ -1,0 +1,6 @@
+"""Constants used throughout the CDF library."""
+
+CONFIG_FILE_NAME = "cdf"
+CONTEXT_PARAM_NAME = "C"
+DEFAULT_DATA_PACKAGES_DIR = "data_packages"
+DEFAULT_DEPENDENCIES_DIR = "dependencies"
