+++ src/cdf/core/context.py
@@ -1,16 +1,7 @@
 """Context module for dependency injection and configuration management."""
 
-import ast
 import asyncio
-import collections
-import importlib.util
 import inspect
-import io
-import json
-import linecache
-import os
-import re
-import string
 import sys
 import threading
 import typing as t
@@ -18,190 +9,25 @@ from contextvars import ContextVar
 from functools import wraps
 from pathlib import Path
 
-import yaml
 from box import Box
-
-if sys.version_info >= (3, 11):
-    import tomllib
-else:
-    import tomli as tomllib
 
 if sys.version_info >= (3, 9):
     from typing import ParamSpec
 else:
     from typing_extensions import ParamSpec
 
+from cdf.core.configuration import SimpleConfigurationLoader
+from cdf.core.constants import CONTEXT_PARAM_NAME
 
 T = t.TypeVar("T")
 P = ParamSpec("P")
 
-ConfigurationSource = t.Union[
-    str, Path, t.Mapping[str, t.Any], t.Callable[[], "ConfigurationSource"]
-]
-
-_CONTEXT_PARAM_NAME = "C"
 
 __all__ = [
     "Context",
-    "ConfigurationSource",
-    "ConverterBox",
-    "SimpleConfigurationLoader",
     "register_dep",
     "active_context",
-    "add_custom_converter",
-    "_get_converter",
-    "_remove_converter",
 ]
-
-
-def _to_bool(value: str) -> bool:
-    """Convert a string to a boolean value."""
-    return value.lower() in ("true", "yes", "1")
-
-
-def _make_eval_func(type_: t.Type):
-    """Create a function to evaluate a py literal string with type assertion."""
-
-    def _eval(value: str) -> t.Any:
-        v = ast.literal_eval(value)
-        if not isinstance(v, type_):
-            raise ValueError(f"Value is not of type {type_}")
-        return v
-
-    return _eval
-
-
-_CONVERTERS = {
-    "json": json.loads,
-    "int": int,
-    "float": float,
-    "str": str,
-    "bool": _to_bool,
-    "path": os.path.abspath,
-    "dict": _make_eval_func(dict),
-    "list": _make_eval_func(list),
-    "tuple": _make_eval_func(tuple),
-    "set": _make_eval_func(set),
-    "resolve": None,
-}
-"""Converters for configuration values."""
-
-_CONVERTER_PATTERN = re.compile(r"@(\w+) ", re.IGNORECASE)
-"""Pattern to match converters in a string."""
-
-
-def add_custom_converter(name: str, converter: t.Callable[[str], t.Any]) -> None:
-    """Add a custom converter to the configuration system."""
-    if name in _CONVERTERS:
-        raise ValueError(f"Converter {name} already exists.")
-    _CONVERTERS[name] = converter
-
-
-def _get_converter(name: str) -> t.Callable[[str], t.Any]:
-    """Get a custom converter from the configuration system."""
-    return _CONVERTERS[name]
-
-
-def _remove_converter(name: str) -> None:
-    """Remove a custom converter from the configuration system."""
-    if name not in _CONVERTERS:
-        raise ValueError(f"Converter {name} does not exist.")
-    del _CONVERTERS[name]
-
-
-def _expand_env_vars(template: str, **env_overrides: t.Any) -> str:
-    """Resolve environment variables in the format ${VAR} or $VAR."""
-    return string.Template(template).substitute(env_overrides, **os.environ)
-
-
-def _read_config_file(
-    path: t.Union[str, Path],
-    mode: str = "r",
-    parser: t.Callable[[str], t.Any] = json.loads,
-    **env_overrides: t.Any,
-) -> t.Any:
-    """Read a file from the given path and parse it using the specified parser."""
-    with open(path, mode=mode) as f:
-        rendered = _expand_env_vars(f.read(), **env_overrides)
-    return parser(rendered)
-
-
-class ConverterBox(Box):
-    """Box that applies @ converters to configuration values."""
-
-    def __getitem__(self, item: t.Any, _ignore_default: bool = False) -> t.Any:
-        value = super().__getitem__(item, _ignore_default)
-        if isinstance(value, str):
-            return self._apply_converters(value)
-        return value
-
-    def values(self) -> t.ValuesView[t.Any]:  # type: ignore
-        return t.cast(
-            t.ValuesView[t.Any], map(self._apply_converters, super().values())
-        )
-
-    def _apply_converters(self, data: t.Any) -> t.Any:
-        """Apply converters to a configuration value.
-
-        Converters are prefixed with @. The following default converters are supported:
-        - json: Convert to JSON object
-        - int: Convert to integer
-        - float: Convert to float
-        - str: Convert to string
-        - bool: Convert to boolean
-        - path: Convert to absolute path
-        - dict: Convert to dictionary
-        - list: Convert to list
-        - tuple: Convert to tuple
-        - set: Convert to set
-        - resolve: Resolve value from partial configuration
-
-        Args:
-            data: Configuration value to apply converters to.
-
-        Raises:
-            ValueError: If an unknown converter is used or if a conversion fails.
-
-        Returns:
-            Converted configuration value.
-        """
-        if not isinstance(data, str):
-            raise ValueError("Value must be a string")
-        data = _expand_env_vars(data)
-        converters = _CONVERTER_PATTERN.findall(data)
-        if len(converters) == 0:
-            return data
-        base_v = _CONVERTER_PATTERN.sub("", data).lstrip()
-        if not base_v:
-            return None
-        transformed_v = base_v
-        for converter in reversed(converters):
-            try:
-                if converter.lower() == "resolve":
-                    try:
-                        transformed_v = self[transformed_v]
-                    except KeyError as e:
-                        raise ValueError(f"Key not found in resolver: {e}") from e
-                else:
-                    transformed_v = _CONVERTERS[converter.lower()](transformed_v)
-            except KeyError as e:
-                raise ValueError(f"Unknown converter: {converter}") from e
-            except Exception as e:
-                raise ValueError(f"Failed to convert value: {e}") from e
-        return transformed_v
-
-
-def _merge_configs(*configs: Box) -> Box:
-    """Combine multiple configuration Boxes using merge_update."""
-    merged = ConverterBox(box_dots=True)
-    for config in configs:
-        merged.merge_update(config)
-    return merged
-
-
-def _scope_configs(*configs: Box) -> Box:
-    """Combine multiple configuration Boxes via ChainMap to provide scope-based resolution."""
-    return ConverterBox(collections.ChainMap(*configs), box_dots=True)
 
 
 class DependencyCycleError(RuntimeError):
@@ -210,84 +36,6 @@ class DependencyCycleError(RuntimeError):
 
 class DependencyNotFoundError(KeyError):
     """Raised when a dependency is not found in the context."""
-
-
-class SimpleConfigurationLoader:
-    """Loads configuration from multiple sources and merges them using a resolution strategy."""
-
-    def __init__(
-        self,
-        *sources: ConfigurationSource,
-        resolution_strategy: t.Literal["merge", "scope"] = "merge",
-        include_env: bool = True,
-    ) -> None:
-        """Initialize the configuration loader with given sources.
-
-        Args:
-            sources: Configuration sources to load.
-            resolution_strategy: Strategy to use for merging configurations
-            - "merge": Merge all configurations into a single Box
-            - "scope": Combine configurations into a ChainMap for scope-based resolution
-            include_env: Whether to include environment variables
-
-        Raises:
-            ValueError: If an unsupported resolution strategy is provided
-        """
-        if resolution_strategy not in ("merge", "scope"):
-            raise ValueError(f"Unsupported resolution strategy: {resolution_strategy}")
-        self.sources = sources
-        if include_env:
-            self.sources += (dict(os.environ),)
-        self._config = None
-        self._resolver = (
-            _merge_configs if resolution_strategy == "merge" else _scope_configs
-        )
-
-    def add_source(self, source: ConfigurationSource) -> t.Mapping[str, t.Any]:
-        """Add a configuration source to the loader.
-
-        Args:
-            source: Configuration source to add.
-        """
-        self.sources += (source,)
-        return self.load()
-
-    def load(self) -> Box:
-        """Load and merge configurations from all sources."""
-        configs = [Box(self._load(source), box_dots=True) for source in self.sources]
-        self._config = self._resolver(*configs)
-        return self._config
-
-    @staticmethod
-    def _load(source: ConfigurationSource) -> t.Mapping[str, t.Any]:
-        """Load configuration from a single source.
-
-        Args:
-            source: Configuration source to load.
-
-        Returns:
-            Configuration as a dictionary.
-        """
-        if callable(source):
-            return SimpleConfigurationLoader._load(source())
-        elif isinstance(source, dict):
-            return source
-        elif isinstance(source, (str, Path)):
-            path = Path(source)
-            if not path.exists():
-                return {}
-            if path.suffix == ".json":
-                return _read_config_file(path, parser=json.loads)
-            elif path.suffix in (".yaml", ".yml"):
-                return _read_config_file(
-                    path, parser=lambda s: yaml.safe_load(io.StringIO(s))
-                )
-            elif path.suffix == ".toml":
-                return _read_config_file(path, parser=tomllib.loads)
-            else:
-                raise ValueError(f"Unsupported file format: {path.suffix}")
-        else:
-            raise TypeError(f"Invalid config source: {source}")
 
 
 class Context(t.MutableMapping[str, t.Any]):
@@ -534,7 +282,7 @@ class Context(t.MutableMapping[str, t.Any]):
             bound_args = sig.bind_partial(*args, **kwargs)
             for name, _ in sig.parameters.items():
                 if name not in bound_args.arguments:
-                    if name == _CONTEXT_PARAM_NAME:
+                    if name == CONTEXT_PARAM_NAME:
                         bound_args.arguments[name] = self
                     elif name in self:
                         bound_args.arguments[name] = self.get(name)
