+++ src/cdf/core/project.py
@@ -1,20 +1,41 @@
+"""Core classes for managing data packages and projects."""
+
+import collections
+import importlib.util
+import sys
+import typing as t
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Union
 
-import tomli
-import yaml
+from cdf.core.configuration import ConverterBox, SimpleConfigurationLoader
+from cdf.core.constants import (
+    CONFIG_FILE_NAME,
+    DEFAULT_DATA_PACKAGES_DIR,
+    DEFAULT_DEPENDENCIES_DIR,
+)
+from cdf.core.context import Context
 
-from .context import Context, SimpleConfigurationLoader
+PathType = t.Union[str, Path]
 
-# Type definitions
-Configuration = Dict[str, Any]
-PathType = Union[str, Path]
+
+def _load_module_from_path(path: Path) -> t.Dict[str, t.Any]:
+    """Load a Python module from a file path."""
+    module_name = path.stem
+    spec = importlib.util.spec_from_file_location(module_name, path)
+    if spec is None or spec.loader is None:
+        raise ImportError(f"Could not load module from path: {path}")
+    module = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(module)
+    return module.__dict__
 
 
 class DataPackage:
-    """Represents a data package with its own context, configurations, and processing logic."""
+    """Represents a data package with its own context and processing logic."""
 
-    def __init__(self, package_path: PathType, parent_context: Context) -> None:
+    def __init__(
+        self,
+        package_path: PathType,
+        parent_context: Context,
+    ) -> None:
         """Initialize the data package.
 
         Args:
@@ -24,61 +45,36 @@ class DataPackage:
         self.package_path = Path(package_path)
         self.name = self.package_path.name
         self.parent_context = parent_context
-        self.context = self._create_context()
-        self.config = self.context.config
 
-        # Load additional configurations or settings as needed
-        self._load_pyproject()
-        self._load_metadata()
+        self.context = self._create_context()
+        self._load_dependencies()
 
     def _create_context(self) -> Context:
         """Create a context for the data package, inheriting from the parent context."""
-        # Load configurations from package-specific sources
-        loader = SimpleConfigurationLoader(
-            self.package_path / "config.yaml", include_env=True
+        return Context(
+            loader=SimpleConfigurationLoader.from_name(
+                CONFIG_FILE_NAME, search_path=self.package_path
+            ),
+            namespace=self.name,
+            parent=self.parent_context,
         )
-        return Context(loader=loader, namespace=self.name, parent=self.parent_context)
 
-    def _load_pyproject(self) -> None:
-        """Load the pyproject.toml file to get package requirements and settings."""
-        pyproject_path = self.package_path / "pyproject.toml"
-        if pyproject_path.exists():
-            with open(pyproject_path, "rb") as f:
-                self.pyproject = tomli.load(f)
-            # Process the pyproject content as needed
-        else:
-            self.pyproject = {}
+    def _load_dependencies(self) -> None:
+        """Load dependencies from Python files in the 'dependencies' directory."""
+        dependencies_dir = self.package_path / "dependencies"
+        if dependencies_dir.exists():
+            sys.path.insert(0, str(dependencies_dir))
+            for py_file in dependencies_dir.glob("*.py"):
+                _load_module_from_path(py_file)
+            sys.path.pop(0)
 
-    def _load_metadata(self) -> None:
-        """Load metadata such as datasets produced and PII indicators."""
-        metadata_path = self.package_path / "metadata.yaml"
-        if metadata_path.exists():
-            with open(metadata_path, "r") as f:
-                self.metadata = yaml.safe_load(f)
-            # Process metadata as needed
-        else:
-            self.metadata = {}
-
-    def get_datasets(self) -> List[str]:
-        """Return the list of datasets produced by this data package."""
-        return self.metadata.get("datasets", [])
-
-    def setup_extraction(self) -> None:
-        """Set up extraction using dlt or other specified tools."""
-        # Implement extraction setup logic, focusing on dlt
-        pass
-
-    def setup_transformation(self) -> None:
-        """Set up transformations using sqlmesh or other specified tools."""
-        # Implement transformation setup logic, focusing on sqlmesh
-        pass
-
-    def run(self) -> None:
-        """Run the data package processing."""
-        self.setup_extraction()
-        self.setup_transformation()
-        # Implement the processing logic
-        pass
+    @property
+    def config(self) -> t.Dict[str, t.Any]:
+        """Get the configuration for the data package."""
+        return ConverterBox(
+            collections.ChainMap(self.context.config, self.parent_context.config),
+            box_dots=True,
+        )
 
 
 class Project:
@@ -92,29 +88,44 @@ class Project:
         """
         self.project_path = Path(project_path)
         self.name = self.project_path.name
+
         self.context = self._create_context()
-        self.config = self.context.config
-        self.data_packages: Dict[str, DataPackage] = {}
+        self._load_dependencies()
+
+        self.data_packages: t.Dict[str, DataPackage] = {}
         self._discover_data_packages()
 
     def _create_context(self) -> Context:
         """Create a context for the project."""
-        # Load configurations from project-specific sources
-        loader = SimpleConfigurationLoader(
-            self.project_path / "config.yaml", include_env=True
+        return Context(
+            loader=SimpleConfigurationLoader.from_name(
+                CONFIG_FILE_NAME, search_path=self.project_path
+            ),
+            namespace=self.name,
         )
-        return Context(loader=loader, namespace=self.name)
+
+    def _load_dependencies(self) -> None:
+        """Load dependencies from Python files in the 'dependencies' directory."""
+        dependencies_dir = self.project_path / DEFAULT_DEPENDENCIES_DIR
+        if dependencies_dir.exists():
+            sys.path.insert(0, str(dependencies_dir))
+            for py_file in dependencies_dir.glob("*.py"):
+                _load_module_from_path(py_file)
+            sys.path.pop(0)
 
     def _discover_data_packages(self) -> None:
         """Discover and load data packages within the project."""
-        packages_dir = self.project_path / "data_packages"
+        data_packages_dir = self.context.config.get(
+            "data_packages_dir", DEFAULT_DATA_PACKAGES_DIR
+        )
+        packages_dir = self.project_path / data_packages_dir
         if packages_dir.exists():
             for package_dir in packages_dir.iterdir():
                 if package_dir.is_dir():
                     data_package = DataPackage(package_dir, parent_context=self.context)
                     self.data_packages[data_package.name] = data_package
 
-    def get_data_package(self, name: str) -> Optional[DataPackage]:
+    def get_data_package(self, name: str) -> t.Optional[DataPackage]:
         """Get a data package by name.
 
         Args:
@@ -125,13 +136,14 @@ class Project:
         """
         return self.data_packages.get(name)
 
-    def run(self) -> None:
-        """Run all data packages in the project."""
-        for data_package in self.data_packages.values():
-            data_package.run()
+    @property
+    def config(self) -> t.Dict[str, t.Any]:
+        """Get the top-level configuration for the project."""
+        return self.context.config
 
 
-# Example usage
 if __name__ == "__main__":
-    project = Project("/path/to/project")
-    project.run()
+    project = Project(".")
+    print(project.data_packages)
+    print(project.context.config["z"])
+    print(project.data_packages["synthetic"].config["z"])
